move exception into correct package
