+++ src/main/java/com/codecool/spaceship/model/exception/ShipNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codecool.spaceship.model;
+package com.codecool.spaceship.model.exception;
 
 public class ShipNotFoundException extends  Exception{
     public ShipNotFoundException() {
